added  view models

+++ composeApp/src/commonMain/kotlin/org/connect/fitconnect/domain/workout/Set.kt
@@ -2,8 +2,8 @@ package org.connect.fitconnect.domain.workout
 
 data class Set(
     val id: Long,
-    val workout_id: Long,
-    val exercise_id: Long,
+    val workout_id: Int,
+    val exercise_id: Int,
     val weight: Double?,
-    val reps: Long
+    val reps: Int
 )
